refactor: 小重构
