+++ Plain Craft Launcher 2.sln
@@ -51,22 +51,22 @@ Global
 		{4499F2D2-18EF-4238-B1F0-7BD547AFF925}.ReleaseUpdate|x64.Build.0 = ReleaseUpdate|x64
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|Any CPU.ActiveCfg = Debug|Any CPU
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|Any CPU.Build.0 = Debug|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|ARM64.ActiveCfg = Debug|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|ARM64.Build.0 = Debug|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|x64.ActiveCfg = Debug|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|x64.Build.0 = Debug|Any CPU
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|ARM64.ActiveCfg = Debug|ARM64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|ARM64.Build.0 = Debug|ARM64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|x64.ActiveCfg = Debug|x64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Beta|x64.Build.0 = Debug|x64
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|x64.Build.0 = Debug|Any CPU
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|ARM64.Build.0 = Debug|ARM64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|x64.ActiveCfg = Debug|x64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Debug|x64.Build.0 = Debug|x64
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|ARM64.Build.0 = Release|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|x64.ActiveCfg = Release|Any CPU
-		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|x64.Build.0 = Release|Any CPU
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|ARM64.ActiveCfg = Release|ARM64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|ARM64.Build.0 = Release|ARM64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|x64.ActiveCfg = Release|x64
+		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.Release|x64.Build.0 = Release|x64
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.ReleaseUpdate|Any CPU.ActiveCfg = Release|Any CPU
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.ReleaseUpdate|Any CPU.Build.0 = Release|Any CPU
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.ReleaseUpdate|ARM64.ActiveCfg = Release|Any CPU
@@ -75,22 +75,22 @@ Global
 		{CECFEF77-5371-4C0B-B40F-A97CF6E0712B}.ReleaseUpdate|x64.Build.0 = Release|Any CPU
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|Any CPU.ActiveCfg = Debug|Any CPU
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|Any CPU.Build.0 = Debug|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|ARM64.ActiveCfg = Debug|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|ARM64.Build.0 = Debug|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|x64.ActiveCfg = Debug|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|x64.Build.0 = Debug|Any CPU
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|ARM64.ActiveCfg = Debug|ARM64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|ARM64.Build.0 = Debug|ARM64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|x64.ActiveCfg = Debug|x64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Beta|x64.Build.0 = Debug|x64
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|x64.Build.0 = Debug|Any CPU
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|ARM64.Build.0 = Debug|ARM64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|x64.ActiveCfg = Debug|x64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Debug|x64.Build.0 = Debug|x64
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|ARM64.Build.0 = Release|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|x64.ActiveCfg = Release|Any CPU
-		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|x64.Build.0 = Release|Any CPU
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|ARM64.ActiveCfg = Release|ARM64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|ARM64.Build.0 = Release|ARM64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|x64.ActiveCfg = Release|x64
+		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.Release|x64.Build.0 = Release|x64
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.ReleaseUpdate|Any CPU.ActiveCfg = Release|Any CPU
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.ReleaseUpdate|Any CPU.Build.0 = Release|Any CPU
 		{A0C2209D-64FB-4C11-9459-8E86304B6F94}.ReleaseUpdate|ARM64.ActiveCfg = Release|Any CPU
