+++ Plain Craft Launcher 2/Pages/PageSetup/PageSetupJava.xaml
@@ -6,7 +6,20 @@
     mc:Ignorable="d" x:Class="PageSetupJava"
     PanScroll="{Binding ElementName=PanBack}">
     <local:MyScrollViewer VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Disabled" x:Name="PanBack">
-        <StackPanel x:Name="PanMain" Margin="25,25,25,10">
-        </StackPanel>
+        <Grid>
+            <StackPanel x:Name="PanMain" Margin="25,25,25,10">
+                <local:MyCard Margin="0,0,0,15">
+                    <StackPanel Margin="8" Orientation="Horizontal">
+                        <local:MyIconTextButton x:Name="BtnRefresh" Text="刷新" LogoScale="0.9" Logo="M875.52 148.48C783.36 56.32 655.36 0 512 0 291.84 0 107.52 138.24 30.72 332.8l122.88 46.08C204.8 230.4 348.16 128 512 128c107.52 0 199.68 40.96 271.36 112.64L640 384h384V0L875.52 148.48zM512 896c-107.52 0-199.68-40.96-271.36-112.64L384 640H0v384l148.48-148.48C240.64 967.68 368.64 1024 512 1024c220.16 0 404.48-138.24 481.28-332.8L870.4 645.12C819.2 793.6 675.84 896 512 896z"/>
+                    </StackPanel>
+                </local:MyCard>
+                <local:MyCard Margin="0,0,0,15">
+                    <StackPanel Margin="15" Orientation="Vertical" x:Name="PanContent"/>
+                </local:MyCard>
+            </StackPanel>
+            <local:MyCard x:Name="CardLoad" HorizontalAlignment="Center" VerticalAlignment="Center">
+                <local:MyLoading x:Name="PanLoad" Text="正在加载 Java 列表" Margin="12"/>
+            </local:MyCard>
+        </Grid>
     </local:MyScrollViewer>
 </local:MyPageRight>
